15.1.Converting HTML to JSX(Section 2: Building Content with JSX) One thing that I see very frequently is very frequently see people who put the return keyword and then
they put the first opening tag of the GSX block on the next line down.

If you do it's right here it's going to result in an air as soon as JavaScript sees that return message

it says.

OK I'm just going to return and there's nothing after the return keyword.

So the function decides to return a value of undefined.

So any time that you return some amount of GSX you have to make sure that the opening tag is on the

same line as the return statement.

A better way to handle this in a way that's a little bit more following in convention in the community

would be any time that you have a multi line segment of GSX like this.

You will instead but the acts on the next line down you evil.

Indent it and then you will wrap it inside of parentheses.

Read the opening parentheses is on the same line as the return statement in the closing parentheses

is after the end of the GSX like so.

So this right here is a very common convention and we're going to be using this convention very frequently

inside this course.

Anytime we have a multi-line GSX block as we do right here OK so everything you see on the screen is

valid so I'm going to save this.

You'll notice that my code formatter automatically clicked kicked in there again just to put a new line

around the label.

That's totally fine till equivalent.

So now I'm going to flip over to my browser and see how this appears in the browser at localhost 3000.

All right so I could flip back over and you'll see that we immediately get a nasty error message here.

You can read the error message if you want to but I'll tell you what it means.