16.Inline Styling with JSX 17.Converting Styling to JSX Format(Section 2: Building Content with JSX)
In the last section we saw a little error message when we took some e-mail and tried to use it inside

of our Riadh component.

So in this video we're going to start to learn some of the very specific differences between GSX and

HMO.

Right.

I got a quick little list right here in general all of the existing knowledge you have around each female

is going to apply directly over to GSX just a handful of tiny differences.

The first difference I want you to be aware of is how we add custom styling to an individual element

as we currently are or at least attempting to on the button element right here.

So we're trying to apply some custom inline styling So that's the first thing that is different between

GSX and HMO.

Let's take a look at another diagram that's going to help you understand how we can take the HTL version

and convert it over to GSX.

OK.

So in this example we've got a div with an inline style of background color red in order to apply that

to J A GSX elements.

We're going to change the syntax just a little bit.

We're still going to say style equals.

But then after that we're going to place two curly braces in a row the outer curly brace so that one

right there indicates that we want to reference a javascript variable inside of our GSX.

Then the second curly brace right there is meant to indicate a javascript object.

So when we do styling with GSX rather than providing a simple string like so we're going to instead

provide a javascript object where all the keys of the object reference a different property that we

want to style and the values indicate the value for that particular styling.

The only other thing that you need to be aware of is that if we ever are trying to apply a style property

name like say background Dasch color we're going to remove that dash and then capitalize the next letter

over.

So anytime that you've got any styling any styling whatsoever that has some word dash another you'll

remove that dash and then capitalize the next character like so.

Now if this was simply say a style meant to customize the background and we wanted say a background

art said let's not do background let's do a border that's a better example.

So if we did say like a border of one picks solid red like so that would translate down to simply Forder

and then inside of a string one picks solid red.

So in general the conversion here is pretty straightforward.

Just remove the double quotes.

Replace some with curly braces and then if you ever have the kind of compound property name remove the

dash into a capitalization.

So with those rules in mind I would love if you took a shot at updating the style on our button tag

right here.

I'm going to take a quick pause right here.

I want you to take a look at the style property and try to convert it over to equivalent GSX.

Now don't spend too much time on this because I haven't even really shown you a good example yet.

Besides this little diagram over here just give it a quick shot after you make a change.

You can save this file and then flip back over to your re-act application and you want to see this error

message go away.

If you see a new error message that's totally fine.

Well fix it up either way in the next video.

So give it a shot and I'll see you in just a minute.

All right.

So did you have any success in converting the style right here if you did it.

Don't sweat it because we're going to get a tremendous amount of practice with GSX over time throughout

this entire course.

So I got to go ahead and convert the inline style right here to the appropriate GSX format instead.

To do so I'm going to first copy the entire button tag and put it on the line right underneath it just

so you can see the before and after practice.

All right so I'm going to first begin by taking off the double quote over here and replacing it with

two curly braces.

Then I'll do the same thing on the right hand side as well.

Next up I'm going to remove the semi-colons.

Inside there and then separate each of the key value properties with the string.

Remember this inner curly brace right here represents a javascript object.

So we need to make sure that it has the appropriate syntax for a javascript object.

Next up I'm going to look at the different property names and find any location where we have a compound

property style.

So that would be right here we have background dash color then remove the dash and capitalized letter

right after it.

And then finally the very last step in this one is a little bit confusing because I didn't very directly

mention it.

We need to make sure that these values are wrapped in a string so I can wrap it with a single quote

like the OK so that's it.

That's how we would update that H-2 e-mail inline style to be the equivalent GSX.

Now I can save this Oliv. both buttons and here just to make sure that they both work.

Actually I shouldn't say that this first button is still going to continue to throw an error or something

to delete that one and then I'll save this file and I'll flip back on my browser and see how we're doing.

Perfect.

So I now see the button is blue with white text.

Now one thing I want to address very quickly.

You might have noticed that for these values right here of blue and white I used single quotes whereas

for all the GSX I use double quotes instead.

So this is a little bit of community convention.

Well have community convention and half requirement of GSX.

First off which sex we are supposed to use double quotes anytime that we want to indicate a string.

So in all these locations here like label name name text we are supposed to use double quotes but then

for any non GSX property by convention in the javascript community traditionally we make use of single

quotes but there are certainly tons of people tons of Engineers tons of companies that prefer to use

double quotes everywhere instead.

So you can do it either way.

It's totally up to you.

At the end of the day there's no difference right here.

If I put double quotes instead of single I just want to highlight that you can make your own decision

there into whichever you prefer.

Now I'm going to be following the convention of using double quotes for all of my GSX properties and

then single quotes everywhere else.

But again do as you please.

OK.

This looks pretty good.