18.Class vs ClassName(Section 2: Building Content with JSX)
In this section we're going to talk about the next tiny little difference between GSX and HMO this difference

is very small so it should get through this video pretty darn quickly.

The rule is anywhere in each female where we assign a class whenever we move over to GSX we're going

to use a slightly slightly different syntax.

So let me show you an example.

Back over inside of our index such as file our label has a class of label right here.

Technically we're not supposed to use the key word class inside of a GSX element.

Instead we use class name like so.

And that's it.

That's the only change you have to make.

Now you might be curious why do we do that.

What is the difference between just saying class and class name.

I want you to recall back when we were working on code pen a little bit ago we had created and yes 20:15

class.

I'm going to show you a quick example of that right here.

We had written something like class app extends re-act dot component.

So the keyword in there was class that was a javascript keyword.

The reason that we have to use class name instead of class in GSX is essentially just to avoid collisions

with this keyword of class that's it.

We just don't want javascript to get confused and think that we are trying to define a new javascript

class right here.

Now as you just saw our code is able to run in the browser without air.

When we do this.

So this is kind of a throwback to the very old days of re-act where this used to be an issue but these

days javascript and the tools around it are intelligent enough to understand the difference between

a class declaration right here and us using the class attribute inside of some GSX.

So with that in mind it's important to understand that there is some discussion in the reac community

right now around removing this requirement of saying class name.

So at some point in time the future you might see in the documentation something that says oh yeah you

can now get away with entering class instead of class name.

However that's something probably in the future might not even by be alive by the time you watch this

video.

Just be aware that this requirement might be a little bit relaxed.

OK.

So replace that with class name or then save the file.

I'm going to flip back over to my browser.

And it looks like everything still continues to work.

OK.

So not bad.

So let's move on to the next section and talk about the last little difference between GSX and HD email.