19.1.Referencing JS Variables in JSX(Section 2: Building Content with JSX)
In this section we're going to talk about one last little difference between GSX and HMO.

This last little thing is probably the most important feature to understand about GSX because we're

going to be working with this feature quite a bit throughout this course.

So GSX can very easily reference javascript variables.

Essentially that means we can take a javascript variable and easily printed up inside of our GSX block.

Let's flip back over to our code editor and get a quick example of what this would look like.

All right.

So inside of my app function right here or my app component I'm going to create a new variable.

They're going to call it something like button text and I'll give it the text of click me.

Like so.

So now if we want to we could refer to this button text variable inside of our button in place of just

putting in the plain text of submit to do so I would delete that text of submit and then I'll put in

a set of curly braces like so now the curly braces here are not indicating a javascript object.

Instead the just the curly braces are indicating that I want to reference a javascript variable right

here so I can then type out the name of the variable that I want to insert into this button.

I can then save this file flip back over to my application at local hold three thousand and I'll see

that.

Now that button says Click me.

So with this curly very syntax we can very easily reference different javascript variables inside of

our GSX.