20.2.Values JSX Can't Show(Section 2: Building Content with JSX)
We could very easily fix this by putting on a button text dot text.

So now that's going to refer to the text property of this object.

We can save this and we'll see that we're back to working as we were before.

Now the reason that I put a lot of emphasis on that last part saying we can't show an object as text

is because technically we can't show a javascript object as a attribute on an element like this right

here.

Now that you understand the purpose of this clearly braes syntax we can look back at the style property

and get a better idea of what's going on.

So we're really doing here is we're saying we want to reference a javascript variable for this style

property and the value that we're going to pass in is this javascript object.

If we wanted to we could technically cut that thing out create a new variable up here called style assign

the style object to it and then reference that style down here like so.

And that would have the exact same effect.

So clearly we can use javascript objects as long as we're not trying to print them up as text or something

like that inside of our application.

OK.

So that's pretty much it.