21 Finding Forbidden Property Names(Section 2: Building Content with JSX)
In the last section we went over these three differences between GSX and HMO.

Now the Sigerson right here I was kind of not being super truthful with you.

Let me be totally honest here.

It turns out that there actually are a couple other differences but at the end of the day it's kind

of challenging for any one person to sit down and list out all of the exact differences between GSX

NHT male.

So in this section I want to teach you how you can figure out some of the differences between these

two on your own.

Let me show you how I want you to flip back over to your react application inside the browser.

Now at present it looks like everything is running just fine.

We can see our content appear on the screen.

However I want you to open up your browser con..

You can do so by right clicking anywhere and clicking on inspect.

You'll then see a window like this appear either on the bottom of the screen or the right hand side

at the top of this panel find the cancel button when you click on that you're going to see a big red

warning like the one I see right here.

This is a warning coming from the re-act library and it says Invalid dumb property for.

Did you mean email for like I said it turns out that there are a couple other differences between GSX

and H.T. mail at the end of the day.

Many of these differences will not crash your application.

And so sometimes it's going to be a little challenging for you to detect when you've done something

slightly wrong in order to detect when you've made a typo or referred to an invalid property you should

traditionally leave your console open or at least reference it every now and then react is going to

look at all the different properties you're passing to elements inside of your GSX blocks.

And if it ever sees that you are passing in an invalid property it's going to show you a warning message

like the one that you see right here.

So this warning message is essentially telling us that our label right here has an attribute name of

four and doesn't want to deal with that.

It wants us to replace it with H T M L for the capital F like so.

Why is that.

Well it's a very similar reason to the class name when we put in for the fear was way back in the earlier

days of re-act that some different browsers or some different tooling might incorrectly see this.

For word right here an incorrectly interpreted as a traditional javascript for loop like so.

So that was the fear as I said earlier not really an issue as you just saw a moment ago our application

works in the browser just fine without error.

Nonetheless the requirement is still encoded in the re-act library.

So we have to replace it with H.T. mail for now if I do that I can save this flip back over and I'll

see that that warning message just went away.

So what's the lesson here lesson is in general as a web developer it's good practice to keep your counsel

open.

Anytime that you see a warning message from react like that something it says did you mean blank.

Usually it's going to mean that you accidentally put together a invalid property name somewhere inside

of your GSX OK.

So with that lesson in mind let's take a quick pause here and continue in the next video.