24.Three Tenets of Components 25.Application Overview(Section 3: Communicating with Props)
All right my friends I think that we understand the basics of GSX pretty well.

So in this video we're going to start to move on to our next big topic.

Now I'm really excited for this section and the next couple of videos because this is where we get to

start taking the training wheels off and start to write a lot more code and move more quickly.

So let's talk about what we're going to be covering in this video in the next couple.

All right.

Here we go.

So three big important concepts for us to understand component nesting reusability and configuration.

These are all tenants of the re-act ecosystem and they're going to govern how we design re-act applications.

Let's talk about each one of these in turn and we'll go through a couple of quick code examples.

So first off component nesting so far out all the code you and I have written we've only been creating

one component at a time.

But for a real re-act application we usually make many different components when we make many different

components.

We then nest them or essentially place them in each other to make an actual usable re-act application.

So we're going to very quickly start to look at a couple examples of how we nece different components

inside of each other.

The next thing we're going to start to discuss is component reusability.

I want you to think about a typical web application that you might use every single day.

Chances are you use a web application that has some buttons or text fields or menus inside of it.

And I got a quick question for you.

Do any of those buttons really look different from each other.

Well chances are they don't just about every button inside of a single re-act application is always

going to have some very similar and identical functionality.

It's going to show some text and then do something when a user clicks on it.

So rather than making a separate component or essentially write out some different code for every single

button or every single menu or every single text input that you might put into an application.

We instead strive to make reusable components or little pieces of code that can be reused several different

times inside of a single application.

And again we'll take a look at a couple of examples on how we might achieve reusability.

Then finally this is very tightly coupled with the usability component configuration.

So I just gave you an example of a button.

Yeah.

Most buttons look very similar but there's little tiny pieces about every single button that need to

be customized in some very specific way.

For example one button might need to have the text submit and another button might need to have the

text click on me and both those buttons might need to do different things.

So we need to make sure that when we create different components we have the ability to somehow customize

them when they are created.

All right.

So these are the three big topics that we're going to start to dive into.

And like I said I'm really excited because this is where things start to get really interesting.

So let's take a quick pause right here.

When we come back the next section we're going to start talking about nesting re-usability and configuration.

In the last section I gave you some fancy terms around components.

Now these terms by themselves don't do a lot of good.

We have to actually write some code to understand what's going on.

So in the section we're going to start to talk about a little application we're going to put together

to get a better idea of what all these different terms mean.

Let me very quickly show you the mockup and then we're going to generate a new project to build this

thing.

OK.

So here's our mockup.

It's really simple rather straightforward.

We're going to essentially make a couple of Riak components that are going to generate what looks like

a list of comments maybe on a blog post or a song or a news article or whatever it might be.

So we're going to want to make sure that we have the ability to display an image a user's name in bold

some time at which the comment was posted.

And then finally the actual comment text underneath it.

Now as you look at this you probably can recognize almost instantly that every little one of these comments

like comment one comment to and comment three.

All look very very similar to each other.

The only difference between them is that they tend to have a little bit of different text different

user names different times and different content for each comment as well.

But as far as the styling goes the spacing all that kind of stuff it generally looks almost identical.

Now rather than just dive in say oh yeah we're going to build a bunch of different components to build

this thing out.

I say that we first begin by taking a kind of naive approach.

By that I mean that we will try to make one single component and in that single component we're going

to put all of the GSX where essentially HCM L that is going to make a list of comments that looks like

this right here after we see some negatives around that approach will then start to refactor it into

separate components.

And so this will give you a much more natural inclination as to why we like to split things out into

reusable components rather than just put down a whole bunch of GSX like a big blob of it.

OK.

So with that in mind let's generate a new project will make the project we'll let our dependencies install.

And all right start writing some code for this in the next video.

So let's start making a new project right now I'm going to flip back over to my terminal where we still

have our GSX project running.

I'm going to shut that thing down by hitting Control-C and then going to go up one directory.

So I'm back in my workspace directory of sorts and then decide if you're going to generate a new re-act

application.

We could absolutely reuse the app that we're working on currently.

That entire scaffold.

But at the end of this course I want you to have a bunch of different projects that you can easily refer

back to if you need to.

So I'm going to run create re-act app and I'm going to get this thing a name of components like so.

OK.

So it's installing my dependencies so let's take a quick pause right here.

When we come back the next section we're going to start putting together this little app.

So see you in just a minute.