29.2.Specifying Images in JSX(Section 3: Communicating with Props)
So now that that is installed I'm going to start my re-act server backup with NPM start OK.

So now we're back to good as we were before and now we can use that Fakhry library inside of our indexed

file.

So to use that library I want to first add an import statement at the very top of the file.

Remember anytime that we want to use some code from another file or a another open source project or

dependency we have to add an import statement.

So I'm going to say import faker from faker like the it's now we can make use of this library to get

a image.

So I'm going to add on that as sarcy which we usually designate in image.

Now in this case we do not have a hard coded image.

You are able to put in here.

So we're not going to put in a set of double quotes like so instead we're going to use that GSX interpellation

where we are going to refer to some javascript variable or the result of some javascript function from

the body of our GSX.

So I'm going to delete those double quotes and I'm going to replace it with that set of double curly

braces like so or single curly braces so then inside of here we're going to reference that Fakhri library.

And then the next part is where things get just a little bit hairy.

Remember back over here inside the documentation there was a section called Image and inside this image

section there was something called Aviator.

So in order to generate a fake image or a fake guitar we're going to call faker dot image.

Ava.

That's how you read these documentation.

If you wanted to make say a fake let's say a fake company name we would call fake or dumb company dot

company name and then call that as a function.

So with that in mind let's flip back over on the say faker that image Eva tar and then are called like

a function by putting on those parentheses like so it's not going to save this file.

I'm going to flip back over to my application in the browser.

I'll see the automatic refresh cookin and then we see the image appear right there.

Awesome.

Now above all the image itself looks pretty darn nice it looks like a real profile image.

The nice thing about faker or maybe the not nice thing.

Let me rephrase that any time you refresh the page the image is going to be automatically regenerated

and you're going to get a new one inserted in there.

So every time that you refresh the page you'll see a new one.

Now sometimes Baker is going to give you a broken image every now and then.

And so you'll see a broken image like this appear.

If that happens you can just refresh the page and he should see a normal image appear again.

OK it says looks fantastic.

Let's take another pause right here and we'll get back on track and figure out how to replicate these

comments and get a couple more down here.