30.Duplicating a Single Component(Section 3: Communicating with Props)
Now that we've got our image in place we're going to take the comment that we generated right here and

we're going to essentially just copy paste it down twice to put two more comments into our application.

Now as you might imagine this is going to turn our app component right here into a gigantic mess of

GSX and it's going to very quickly look absolutely hideous and rather challenging to navigate nonetheless.

Let's go to this.

This process and you're going to very quickly understand why we like to make reusable components.

OK.

So Eugene to duplicate this comment I'm going to find the div with the class name of comment right here

and I'm going to highlight everything down to the closing DIV tag for that comment.

Lexa and then going to copy this thing and I'm going to paste it underneath it two more times.

All right.

So now this thing is a real beast to navigate without a doubt.

Now going to save this thing I'm going to flip back over to the browser and we're going to see that

sure enough.

We do have three distinct comments.

Now of course if we want to we could go through and change the name and the time and the actual blog

content on there as well.

But more are going to react to this in just a moment anyways.

So without a doubt I'm sure you can scroll through this file.

And if I told you go and find the second comment and edit the title.

Well if you're at the very top you kind of got to like Scroll through this thing and try to understand

what the heck you're looking at.

So there's really no two ways about it.

By just duplicating all that GSX inside of a single component we end up in a pretty bad spot and it

tends to be rather challenging for us to make any update to this app component.

So I think that we've probably now feel the pain right like this is definitely not good.

So let's take another quick pause right here.

When we come back the next section we're going to talk about how we're going to really factor this thing

into some separate components and how we're going to customize each of those individual components.

So in quick pause and I'll see you in just a minute.